Score update
Score update.

+++ Framework-2025/SDL_Template/StartScreen.h
@@ -89,7 +89,7 @@ private:
 	Vector2 mSliderKnobEndPos;
 	bool mInOptions;
 
-	int mHighScore;
+	float mHighScore;
 
 public:
 	StartScreen();
